pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "click>=8.2.1",
    "flask>=3.1.1",
    "markdown>=3.8.2",
    "openpyxl>=3.1.5",
    "pytest>=8.4.1",
    "pytest-cov>=6.2.1",
    "pyyaml>=6.0.2",
    "werkzeug>=3.1.3",
]
